# Add sliding window attention bias 