@drisspg 2024-03-05
@desertfire Do you know what the ramifications are of the failing ABI compat test for aotinductor, in theory this should be compatible since I am add kwarg only options to private ATen ops but not sure if there is documentation / detail on the process for doing so