@desertfire 2024-03-07
> @desertfire Do you know what the ramifications are of the failing ABI compat test for aotinductor? In theory this should be compatible since I am adding kwarg only options to private ATen ops but not sure if there is documentation / detail on the process for doing so

We already have a v2 `aoti_torch__scaled_dot_product_flash_attention_v2`  for some historical reason, https://github.com/pytorch/pytorch/blob/b9087f857147da37c122cb70955d791565ff1ed6/torch/csrc/inductor/aoti_torch/shim_common.cpp#L376. Normally, you will need to create a similar `v3` version with additional parameters you are adding here, and change the codegen rule here, https://github.com/pytorch/pytorch/blob/b9087f857147da37c122cb70955d791565ff1ed6/torch/_inductor/ir.py#L4954.

But I have a recent PR touching some relevant pieces, https://github.com/pytorch/pytorch/pull/120513. So things are slightly messy here. If you rebase after my PR, it should work for OSS, but there will be issues for internals. If you think this is the only remaining blocker, I can work out a patch PR to unblock.